pyproject.toml
[project]
name = "llm-dna"
version = "0.2.2"
description = "Extract LLM DNA vectors — low-dimensional representations that capture functional behavior and model evolution."
authors = [{ name = "LLM-DNA Project" }]
license = { file = "LICENSE" }
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "numpy>=1.24.0",
    "torch>=2.4.0",
    "transformers>=4.40.0",
    "accelerate>=0.26.0",
    "huggingface-hub>=0.23.0",
    "datasets>=2.14.0",
    "pandas>=2.0.0",
    "sentence-transformers>=2.2.2",
    "scikit-learn>=1.3.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
    "wonderwords>=2.2.0",
    "openai>=1.0.0",
    "tiktoken>=0.7.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
model_scraping = ["requests>=2.31.0"]

apple = [
    "mlx>=0.10.0; sys_platform == 'darwin' and platform_machine == 'arm64'",
    "mlx-lm>=0.10.0; sys_platform == 'darwin' and platform_machine == 'arm64'",
]

quantization = [
    "bitsandbytes>=0.46.1",
    "autoawq>=0.2.0",
    "auto-gptq>=0.5.0",
    "optimum>=1.16.0",
    "compressed-tensors>=0.1.0",
]

model_families = [
    "mamba-ssm>=1.0.0; sys_platform == 'linux'",
    "timm>=0.9.0",
]

full = [
    "llm-dna[apple]",
    "llm-dna[quantization]",
    "llm-dna[model_families]",
]

vllm = ["vllm>=0.4.0"]

dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]

[project.scripts]
llm-dna = "llm_dna.cli:main"
calc-dna = "llm_dna.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/llm_dna"]

[tool.hatch.build.targets.sdist]
include = [
    "src/llm_dna/",
    "scripts/calc_dna.sh",
    "scripts/calc_dna.py",
    "configs/llm_list.txt",
    "README.md",
    "LICENSE",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-q"
markers = [
    "slow: marks tests as slow (run with --run-slow)",
]
